Feat/grid world

+++ src/noise.rs
@@ -2,6 +2,7 @@ use noise::{Billow, Fbm, NoiseFn, Perlin, RidgedMulti, ScalePoint, Seedable, Tur
 
 // Noise generation using Perlin noise
 
+#[derive(Debug)]
 pub struct NoiseGenerator {
     pub seed: u32,
     pub perlin: Perlin,
@@ -29,16 +30,16 @@ impl NoiseGenerator {
     }
 
     // Basic height map generation using FBM
-    pub fn height_map(&self, x: f64, y: f64) -> f64 {
+    pub fn height_map(&self, x: f32, y: f32) -> f32 {
         let height_noise = ScalePoint::new(self.fbm.clone())
         .set_scale(self.scale);
-        (height_noise.get([x, y]) + 1.0) * 0.5
+        ((height_noise.get([x as f64, y as f64]) + 1.0) * 0.5) as f32
     }
 
     // Temperature noise for biome generation
-    pub fn temperature_noise(&self, x: f64, y: f64) -> f64 {
+    pub fn temperature_noise(&self, x: f32, y: f32) -> f32 {
         let temp_noise = ScalePoint::new(self.perlin.clone().set_seed(self.seed + 1))
             .set_scale(self.scale);
-        (temp_noise.get([x, y]) + 1.0) * 0.5
+        ((temp_noise.get([x as f64, y as f64]) + 1.0) * 0.5) as f32
     }
 }
